FPS Upload
I spent several hours working on developing a working Fire/Ice Gun and it looks like it finally works as intended. If you load up the scene that says "FPS" and play it, you can shoot the floating block and it will turn blue or orange depending on which bullet you used.