it works!
4 types of obsticals are generated in front of the player. the spawner moves with the player so obsticals are allways generated aways away.

these are the obsticales:
fire_1
fire_2
ice_1
ice_2
basically two sizes for each for now

Experiments:
added a simple encloser to the level and made it rotatable by pressing Q and E. They are different colors so you can see it actually rotating.

Working on getting the obsticals rotating with the level aswell.